requirements.txt
# LlamaParse Ingestion Pipeline Requirements
# Install with: pip install -r requirements.txt

# Core LlamaParse and LangChain dependencies
llama-cloud-services>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3

# LlamaIndex for semantic chunking
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data processing
numpy>=1.21.0
pandas>=1.5.0

# Optional: For GPU acceleration (uncomment if you have CUDA)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Utility packages
python-dotenv>=1.0.0
tqdm>=4.64.0

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0